+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 2.2.0 Alpha 1 (XXXX-YY-ZZ)
 
 * `permissions` and `options` support for event listeners.
+* `permissions` and `options` support for template listeners.
 * `wcf\data\TDatabaseObjectOptions` and `wcf\data\TDatabaseObjectPermissions` for database object-bound options and permissions validation.
 * `wcf\system\cache\builder\EventListenerCacheBuilder` returns `wcf\data\event\listener\EventListener` objects instead of data arrays.